add groups: extract duplicate create group logic
